% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date.R
\name{add_date}
\alias{add_date}
\title{Add a date column to a dataframe containing Year and Day columns}
\usage{
add_date(dataframe, is_day = NULL, is_year = NULL, is_hour = NULL)
}
\arguments{
\item{is_day}{A logical vector with a value for every column in dataframe indicating if
the column contains day of year data. If not provided, then will attempt to generate it using
the is_day function}

\item{is_year}{A logical vector with a value for every column in dataframe indicating if
the column contains year data. If not provided, then will attempt to generate it using
the is_year function}

\item{datafrane}{A dataframe to which a date column should be added}
}
\description{
A function which accepts a dataframe, a logical vector indicating which column of
the dataframe contains day data, and a logical vector indicating which column of
the dataframe contains year data. If no logical vectors are provided then they will
be produced using the day_identifier and year_identifier functions.
The day of year and Year data are used to produce a new column in the dataframe named
"date", using the lubridate package. The new dataframe, with appended date column is returned
}
