% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Remove Duplicate and Out-of-Bounds Data}
\usage{
clean_data(
  dataframe,
  upper_limit,
  lower_limit,
  name_pattern,
  output_dir,
  grouping = "date"
)
}
\arguments{
\item{dataframe}{The dataframe containing the data to be cleaned}

\item{upper_limit}{The limit above which data in the identified columns will be converted
to NA}

\item{lower_limit}{The limit below which data in the identified columns will be converted
to NA}

\item{name_pattern}{A string or regular expression which should match all columns,
and only those columns, to be cleaned}

\item{output_dir}{Directory to which a file of all conflicted data shall be written}
}
\description{
A function which takes a dataframe, and a name_pattern to identify columns within the
dataframe. For each column that matches the name_pattern, its values are converted to NA
if they are outwith the upper or lower limits, or if they match 6999 (an error code produced by
Tundra lab equipment). For every date, the data is checked to make sure any duplicate data
is self consistent. Any inconsistent data is written to a new file (output_dir/conflicted_data.csv)
before being converted to NA, except if discarding NA values would lead to consistent data.
All duplicated data is then removed from the dataframe before returning it.
}
