% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_re-compiler_function.R
\name{create_recompiler_function}
\alias{create_recompiler_function}
\title{Re-compile parent file}
\usage{
create_recompiler_function(file_path, test_function, ...)
}
\arguments{
\item{test_function}{The function to be applied to the parent file}

\item{...}{Any further parameters required by the test function}

\item{files}{Either a file path to a list of compiled files, or a list of the compiled
files}
}
\description{
A function designed for troubleshooting. After running list_erroneous_files() a
file (erroneous_files.txt) will be produced containing a list of all compiled files
with the wrong number of columns. create_re-compiler_function takes the file path to
this text document as input, as well as the function used to produce the output (or
conceivable any other function you might wish to apply to the parent files). This
from this function is a function which combines the features of create_file_reader_function()
and the probe_errors() function on the parent files. The output is a function which,
each time it is run will produce a list of the next parent file, and of the output of
the test function (which will be the final compiled file if using the compiler function as
the test function). The main utility of this function is to use it in conjunction with
the R studio debugger to determine where the compiling went wrong. A list of file paths
may be used as input instead of a file path to a text document containing them.
}
