% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year_from_string.R
\name{date_extractor}
\alias{date_extractor}
\title{extract dates and components from within a string}
\usage{
date_extractor(
  string,
  date_format = "all",
  date_day_month = "date",
  breaks = FALSE,
  unique = TRUE
)
}
\arguments{
\item{string}{A character vector that is to be examined for days}

\item{date_format}{which date formats will be checked. One of "all", "ymd" (year-month-day),
"dmy" (day-month-year), or "mdy" (month-day-year). Defaults to all.}

\item{date_day_month}{Should the whole date, only the day, or only the month be returned}

\item{breaks}{Should results be restricted to dates encapsulated by word boundaries}

\item{unique}{Should result return all unique hits or a full list for counting occurrences}
}
\description{
A function which accepts a character vector and attempts to identify dates within the string.
If no dates are found then it returns NULL. Accepted formats include:
year-month-day, day-month-year, or month-day-year, with or without separators between the date
components. date_day_month can be set to either "date", "day", or "month", defining whether the
function shall return the whole date, only the day, or only the month.
A date format can be specified resulting in only dates of this format producing results. If no
format is specified, then only days greater than 12 will be returned as it is otherwise not
possible to identify the day from the month. If searching for date components of an unknown format
the safest technique is to first extract the dates, then extract the day, and only proceed with
finding the month if a day is returned. To further prevent miss-atributation, only years
1980-2029 will be returned. Setting breaks to TRUE means that
only dates isolated from other words are returned, for example "summer20120517" would not produce
a hit but "summer_20120517" or simply "20120517" would. Setting unique to FALSE means all hits are
returned, which is useful for counting them.
}
