% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_identifier.R
\name{day_identifier}
\alias{day_identifier}
\title{Identify a column containing Day data within a dataframe}
\usage{
day_identifier(dataframe)
}
\arguments{
\item{dataframe}{A dataframe which will be analysed to determine if its columns contain day of year data}
}
\description{
A function which analyses every column in dataframe and returns a logical vector where
the value true indicates the corresponding column contains a list of days (1-366).
A column is identified as containing day data if, ignoring the 1st and last quartiles to avoid
being thrown by errors, the data ranges from 30 - 366, as well as containing data that, at least 90\%
of the time increases by exactly 1 between unique values.
}
