% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_sites.R
\name{determine_annex}
\alias{determine_annex}
\title{A function which first checks every row of the dataframe to see if "anne" can be found within
any of its columns. If so that row is marked as belonging to an Annex site. All rows deemed to
be belonging to annex sites in this way have "Annex_" prepended to their site names if they don't
already have so.}
\usage{
determine_annex(dataframe)
}
\description{
A function which first checks every row of the dataframe to see if "anne" can be found within
any of its columns. If so that row is marked as belonging to an Annex site. All rows deemed to
be belonging to annex sites in this way have "Annex_" prepended to their site names if they don't
already have so.
}
