% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_identifiers.R
\name{determine_identifiers}
\alias{determine_identifiers}
\title{Create list of all identifiers found in identifier column of a dataframe}
\usage{
determine_identifiers(dataframe, is_identifier)
}
\arguments{
\item{dataframe}{A dataframe which contains the identifier column to be inspected}

\item{is_identifier}{A logical vector with a value for each column in dataframe indicating
if it contains identifier data}
}
\description{
A function which accepts a dataframe and a logical vector identifying each column
of the dataframe as either an identifier column or not. The function returns a list
of the identifiers found within the column identified by is_identifier. If is_identifier
indicates no identifier column is present then the function returns NULL
}
