% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_type_identifier.R
\name{file_type_identifier}
\alias{file_type_identifier}
\title{Determine if a file is CSV or Tab deliminated}
\usage{
file_type_identifier(file_path)
}
\arguments{
\item{file_path}{A path to the file to be analysed}
}
\description{
A function which takes a file_path to a file and attempts to determine if this file
is either CSV or Tab delimited. It eliminates either possibility if
\enumerate{
\item more than 30\% of rows are predicted to have only 1 column
\item less than 90\% of rows are predicted to have more than 3 columns
\item There are predicted to be more than 5 different row lengths
At each test, if ambiguous, CSV is favoured. If all tests fail, file is predicted to
be CSV
}
}
