% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_header.R
\name{find_header}
\alias{find_header}
\title{Locate header line within CSV or Tab deliminated file}
\usage{
find_header(file_path, sep, multi_header = FALSE)
}
\arguments{
\item{file_path}{A path to the file that is to be checked for a header line}

\item{sep}{A character vector containing the character used to deliminate the file}
}
\description{
A function which accepts a file_path and separator and attempts to Find the position
of the column headers, if any, within that file. It ignores meaningless header names
"...(number)" and "X(number)". A row is assumed to be a header line if columns 2,3,
and 4 do not contain numerical data. up to the first 10 rows are checked. If a header is
found, the number of rows to skip while reading in the file to make the header the first
row read is returned. If no header is found the string "no header" is returned.
}
