% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_treatment_cols.R
\name{find_treatment_cols}
\alias{find_treatment_cols}
\title{Find Treatment Columns}
\usage{
find_treatment_cols(dataframe, output_dir, file_path)
}
\arguments{
\item{dataframe}{A dataframe containing columns with names matching regex "otc|treat" containing
which are to be identified as either otc_treatment or snow_treatment}

\item{output_dir}{The directory that a warning file will be written to if snow and otc treatment
data can not be clearly separated.}

\item{file_path}{A file path for the original data. Used in the output of a warning message if
snow and otc treatment data could not be clearly separated.}
}
\description{
A function which accepts a dataframe and returns it with any treatment columns identified as either
otc_treatment or snow_treatment. If a column is present containing both snow and otc treatment
then the data will be split into otc and snow treatment columns. If a mixed column is present as well
as a dedicated snow or otc column this will cause the function to crash as it is not possible to
determine which data is to be kept. Any time a mixed otc/snow column is found a warning file is
produced in the output_dir containing the file_path of the original data. This is because it is
difficult to determine if "control" should belong to snow or otc.
}
