% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_ID.R
\name{get_plot_ID}
\alias{get_plot_ID}
\title{Get plot ID}
\usage{
get_plot_ID(
  site_name,
  plot_number,
  otc = NA,
  snow = NA,
  plot_df = tundra::plot_names
)
}
\arguments{
\item{site_name}{A character vector providing the name of the site}

\item{plot_number}{A character vector providing the plot number. Must be a string, as some
plot numbers either include or are exclusively letters}

\item{otc}{A character vector indicating the otc treatment applied}

\item{snow}{A character vector indicating the snow treatment applied}

\item{plot_df}{The dataframe containing the plot IDs. Defaults to the plot_names dataframe
stored within the tundra package}
}
\description{
A function which accepts as a minimum, a site name, and plot number, with additional options
for otc and snow treatment. The function looks up the plot_names dataframe stored within the
tundra package, and returns the plot ID which is unique to the provided site/plot characteristics.
If multiple plots could have the provided characteristics then all applicable plot IDs are
returned in one character string separated by /. If no valid plot Ids can be identified then
NA is returned.
}
