% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year_from_string.R
\name{grep_capture}
\alias{grep_capture}
\title{Extract from a regex capture statement}
\usage{
grep_capture(regex, string, target = NULL)
}
\arguments{
\item{regex}{A character string containing a regex expression with capture points}

\item{string}{The character string which is to be searched}

\item{target}{set to the default NULL. If grep_capture is to be vectorised across
a list of strings then target can be used to pass the results from previous searches.
The final result will then be an aggregate of all hits.}
}
\description{
A function which accepts a regex expression and a string to be searched. If the regex
finds a hit within the string, and the regex contains brackets for capturing part of
the hit then the captured text is returned. If there are multiple captured texts then
they are returned individually in a character vector.
}
