% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_id_extractor.R
\name{hobo_id_extractor}
\alias{hobo_id_extractor}
\title{extract hobo id from string}
\usage{
hobo_id_extractor(string, unique = TRUE)
}
\arguments{
\item{string}{A character vector that is to be examined for sites}

\item{unique}{Should result return all unique hits or a full list for counting occurrences}
}
\description{
A function which accepts a character vector and returns all unique instances within that string
that identify a hobo serial number. If no sites are found then it returns NA. Setting unique to FALSE means
all hits are returned, which is useful for counting them. If multiple sites are present within one
string then an error is returned
}
