% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hour_identifier.R
\name{hour_identifier}
\alias{hour_identifier}
\title{Identify which column of a dataframe contains time data}
\usage{
hour_identifier(dataframe, is_day = NULL, is_year = NULL)
}
\arguments{
\item{dataframe}{A dataframe to be checked for the presence of time data}

\item{is_day}{An optional logical vector indicating which column contains the day data.
If not provided it is generated using the day_identifier function}

\item{is_year}{An optional logical vector indicating which column contains the day data.
If not provided it is generated using the year_identifier function}
}
\description{
A function which accepts a dataframe that is to have its columns analysed to determine
if they contain time data or not, as well as optional logical vectors detailing which
columns contain day data and year data. If these are not provided then they will be generated
using the day_identifier and year_identifier functions.
hour identifier determines if a column contains time data by first ensuring 90\% of the
values fit the format XX0 or YXX0, where Y is a digit between 0 and 2, and X is any
numeral. If the appropriate format is found, then the data is checked to see if all
hours (0-23) are present. Only the first column containing time data is identified as
a time column. Only the first 8 columns are investigated. A list is returned. The first
entry of which is a logical vector indicating which column contains time data. The second
entry is the dataframe modified to contain a decimalised fractional day column
}
