% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_witout_target.R
\name{identifier_without_target}
\alias{identifier_without_target}
\title{Identify which column of a dataframe contains identifier data without specifying a specific
identifier}
\usage{
identifier_without_target(dataframe, output_dir, file_path)
}
\arguments{
\item{dataframe}{A dataframe that should be analysed for the presence of an identifier column.}

\item{output_dir}{The directory to which a text file will be written identifying any files
for which multiple possible identifier columns were detected. If this happens, only the first
result is taken.}

\item{file_path}{The path to the file containing the data loaded into dataframe. Needed for
making the text file described in output_dir}
}
\description{
A function to identify the identifier column without looking for a specific identifier number.
The identifier column contains a numerical value that describes the type of data to expect
on that row. A column is determined to be an identifier column if at least 80\% of its values
consist of only either 2 or 3 digits, and contains no more than ten unique values and is not
a day column. Only the first 5 columns are checked for efficiency.
The function returns a logical vector with TRUE/FALSE values for every column in dataframe.
}
