% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is.empty}
\alias{is.empty}
\title{Is a variable NA, Null, or an empty string}
\usage{
is.empty(var)
}
\arguments{
\item{var}{The Variable to be examined}
}
\description{
A simple function returning the answer to the logical statement
is.na(var) | is.null(var) | var == ""
}
