% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_erronous_files.R
\name{list_erronous_files}
\alias{list_erronous_files}
\title{Produce list of miss-compiled files}
\usage{
list_erronous_files(directory, min_cols, max_cols)
}
\arguments{
\item{directory}{The directory containing the compiled files to be examined. Also, where the
list of miss-compiled files will be written to.}

\item{min_cols}{The minimum number of columns a compiled file should have}

\item{max_cols}{The maximum number of columns a compiled file should have}
}
\description{
A function which takes the file path for an output directory for one of the compiler
scripts, as well as max and minimum number of columns those files should have if they
compiled correctly. It then produces a list of the erroneous files in the text document
erroneous_files.txt in the same directory. This list can then be used as input to efficiently
examine the miss-compiled files using the create_file_reader_function function, or the
original compilation may be investigated using the create_recompiler_function function.
}
