% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file.R
\name{load_file}
\alias{load_file}
\title{Load CSV or Tab deliminated files despite asymmetry}
\usage{
load_file(file_path, sep = NULL, split_file = FALSE)
}
\arguments{
\item{file_path}{A path to the file which is to be loaded}

\item{sep}{A character vector of the separator used by the file to distinguish columns}

\item{split_file}{If TRUE, a list of dataframes is returned, each dataframe consisting of all
rows with the same number of columns}
}
\description{
A function to load the data from the file identified in file_path. If no separator is provided
it uses file_type_identifier to predict it. It also uses find_header()
to attempt to locate meaningful header names, and sets the variable skip accordingly (see
find_header() for details). If the header does not cover every column in the file then the
header is ignored. load_file is configured to handle asymmetric files, and to strip white
space from fields. All columns are read as characters to provide consistent data retrieval.
If split_file TRUE, then a list of dataframes is returned, each dataframe consisting of all
rows with the same number of columns
}
