% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_specific_row_length.R
\name{load_specific_row_length}
\alias{load_specific_row_length}
\title{A function to read a file and extract only the rows with a length corresponding to
row_length. This function is used by load_file to produce a list of dataframes, where
each element of the list is a dataframe of all rows with the same row_length.}
\usage{
load_specific_row_length(file_path, row_lengths, sep = ",", row_length)
}
\arguments{
\item{file_path}{The file path to the tabular data which is to be loaded}

\item{row_lengths}{The number of columns in every line of the file. Can be produced using the
count.fields function}

\item{sep}{A character vector detailing the delimiter for the file: either "," or ""}

\item{row_length}{The number of columns a row should have to be included in the outputed
dataframe}
}
\description{
A function to read a file and extract only the rows with a length corresponding to
row_length. This function is used by load_file to produce a list of dataframes, where
each element of the list is a dataframe of all rows with the same row_length.
}
