% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_lookup.R
\name{lookup}
\alias{lookup}
\title{Generic lookup function}
\usage{
lookup(key_col_name, key, lookup_table)
}
\arguments{
\item{key_col_name}{What is the name of the column that should be searched for the key}

\item{key}{The thing which should uniquely identify the data that is being searched for
such as an ID number}

\item{lookup_table}{The dataframe which should be searched for the data}
}
\description{
A function which accepts a key column name, key, and lookup table, searches the
key_column in the lookup table for the key, and returns a named list containing
all the other column entries that corespond with that key. If more than one hit
is found for the key, or no hits are found, then a list of NAs are returned. The
names for the returned list are taken from the names of the columns in the lookup
table.
}
