% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_treatment.R
\name{lookup_treatment}
\alias{lookup_treatment}
\title{Look up treatments given plot ID}
\usage{
lookup_treatment(id, plot_df = tundra::plot_names)
}
\arguments{
\item{id}{The plot Id to be looked up to find the corresponding treatment data}

\item{plot_df}{The dataframe which will be searched for the treatment data. Defaults
to tundra::plot_names}
}
\description{
A function which accepts a plot_id and returns a named list providing the OTC,
Snow, and Fertaliser treatments. The data is looked up in the tundra::plot_names
dataframe. If the plot_id does not return one, and only one hit, then NA is returned.
}
