% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Determine the Mode of a set of data}
\usage{
Mode(data)
}
\arguments{
\item{data}{A vector containing the data that should have its mode determined}
}
\description{
A function to return the most repeated value in a set i.e. the mode. This function
can be applied to both numeric and categorical data. If there are multiple values
sharing the maximum number of repetitions then the first of these values is returned.
}
