% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_NA.R
\name{non_NA}
\alias{non_NA}
\title{return first value from a vector which isn't NA
A function which accepts a vector and returns the first value which isn't NA.
Most notably used by the clean_data function to retain non NA data where available}
\usage{
non_NA(values)
}
\arguments{
\item{values}{A vector containing the values to be searched for non NA data}
}
\description{
return first value from a vector which isn't NA
A function which accepts a vector and returns the first value which isn't NA.
Most notably used by the clean_data function to retain non NA data where available
}
