% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extractor.R
\name{plot_name_extractor}
\alias{plot_name_extractor}
\title{Extract plot information from a string}
\usage{
plot_name_extractor(string, unique = TRUE, as_null = TRUE)
}
\arguments{
\item{string}{A character vector that is to be examined for sites}

\item{unique}{Should result return all unique hits or a full list for counting occurrences}

\item{as_null}{If TRUE then NULL is returned if there is no hits, which is good for counting.
If False, then NA is returned if no hits, good for working with dataframes}
}
\description{
A function which accepts a character vector and returns a plot name, if present in the string.
Plot names consist of the plot number followed by an abbreviation for the treatment, for example
1C, 11Cc, or 13OTC. This function does not work for the farm or fert sites, or for any data where
only the plot number is present unless the word "plot" directly precedes the plot name.  If no sites
are found then it returns NULL. Setting unique to FALSE means all hits are returned, which is useful
for counting them.
}
