% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_specific_row_length.R
\name{reheader_dataframe}
\alias{reheader_dataframe}
\title{A function which accepts a dataframe as input. It checks the first 10 rows using the
search_dataframe_for_header function, to determine if a header line is found within the
data. If such a header is detected it returns a new dataframe excluding any data above
the header line and with the columns named after the found header line. If no header
is detected it returns the original dataframe unchanged}
\usage{
reheader_dataframe(dataframe)
}
\description{
A function which accepts a dataframe as input. It checks the first 10 rows using the
search_dataframe_for_header function, to determine if a header line is found within the
data. If such a header is detected it returns a new dataframe excluding any data above
the header line and with the columns named after the found header line. If no header
is detected it returns the original dataframe unchanged
}
