% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reroute_file_paths.R
\name{reroute_file_paths}
\alias{reroute_file_paths}
\title{Change the route dir of a file path}
\usage{
reroute_file_paths(path_to_route, file_paths)
}
\arguments{
\item{path_to_route}{The file path to the first common directory to both the
old and new file paths}

\item{file_paths}{All old file paths which have to be modified}
}
\description{
A vectorised function which accepts a list of file paths, and a path_to_route
detailing where the new location of the route directory. The function returns a
list of file paths pointing to the provided path_to_route. For example, if the
file path "C:/old/file/path/AF_database/things/stuff/data.csv" is to be updated
to "C:/new/path/to/file/AF_database/things/stuff/data.csv" then the path_to_route
would be "C:/new/path/to/file/AF_database". If the original file path does not
include the last directory of the path_to_route then the old file path is returned
unchanged and a warning is issued.
}
