% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_compile.R
\name{rough_compile}
\alias{rough_compile}
\title{Combine dataframes from file}
\usage{
rough_compile(file_list)
}
\description{
This function accepts as input a list of file paths and outputs a dataframe of
All of the associated files combined together, with any columns not pressent for
a given file populated NA. The function also produces a column called file, which
contains the file path to the original data. This function uses the load_file
function to automatically detect the header lines of the given files. This will
work quite reliably so long as the files are at least 5 columns wide.
@param file_list A list of files to be loaded and merged into one dataframe
}
