% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dataframe.R
\name{save_dataframe}
\alias{save_dataframe}
\title{Save a dataframe with standard naming format, avoiding overwrite potential}
\usage{
save_dataframe(dataframe, output_dir, name)
}
\arguments{
\item{dataframe}{A dataframe to be saved}

\item{output_dir}{The directory to receive the saved file}

\item{name}{The name for the file, which will be augmented with year and day data}

\item{is_year}{A logical vector indicating which column of the dataframe contains year
data. If absent, will be generated using function year_identifier}

\item{is_day}{A logical vector indicating which column of the dataframe contains day
data. If absent, will be generated using function day_identifier}
}
\description{
A function which takes a dataframe and writes it to file. It names the file with
a name provided in the name parameter in the format <name>_FIRST-YEAR-LAST-YEAR.csv, if
the file spans multiple years, or <name>_YEAR_FIRST-DAY-LAST-DAY.csv if it doesn't.
If a file with the same name already exists, an iterator is added to the end. Files
are written to the output directory. Year and Day data should be identified by logical vectors
indicating which columns of the dataframe contain Year and Day of year data. If these
logical vectors are not provided they are generated using the functions year_identifier
and day_identifier.
}
