% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_dataframe_for_header.R
\name{search_dataframe_for_header}
\alias{search_dataframe_for_header}
\title{A function which takes a dataframe as input, and returns the row number where it
has detected a posible header line. A line is considered to be a header if in rows
2,3, and 4 it contains non-numeric, non-blank data. If 2 potential header lines are
found consecutively, the upper line is checked for known meaningless headers (V1-9, X1-9
...1-9), and if so the lower row is returned as header. Otherwise, or if more than 2
potential headers are detected or dataframe has less than 4 columns, then the character
string "no header" is returned.}
\usage{
search_dataframe_for_header(dataframe, search_limit = 10, multi_header = FALSE)
}
\arguments{
\item{dataframe}{The dataframe which is to be checked for the presence of a header line
within its data}

\item{search_limit}{The number of rows from the top to check for a header. Defaults to
10. If the limit is set higher than the number of rows in dataframe then all rows are
checked}
}
\description{
A function which takes a dataframe as input, and returns the row number where it
has detected a posible header line. A line is considered to be a header if in rows
2,3, and 4 it contains non-numeric, non-blank data. If 2 potential header lines are
found consecutively, the upper line is checked for known meaningless headers (V1-9, X1-9
...1-9), and if so the lower row is returned as header. Otherwise, or if more than 2
potential headers are detected or dataframe has less than 4 columns, then the character
string "no header" is returned.
}
