% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_identifier.R
\name{solar_identifier}
\alias{solar_identifier}
\title{Identify which columns of a dataframe contain solar data}
\usage{
solar_identifier(dataframe, file_path, output_dir, W_or_KW = "KW")
}
\arguments{
\item{dataframe}{A dataframe to be inspected for solar data. Must have a date column as its
last column. This can be added using the add_date function}

\item{file_path}{The path to the file from which the data was written. This file path will
be written to output_dir/cant_determine.txt if 80\% or more of the data was from Oct, Feb, Mar}

\item{output_dir}{The directory that the file cant_determine.txt will be written to if
a file is found with 80\% of its data from Oct, Feb, Mar, which can not be ruled out as not
solar data}

\item{W_or_KW}{A character vector of either "W" or "KW" to permit the setting of the units
expected. Defaults to KW/m2 if no value is given}
}
\description{
A function which accepts a dataframe and returns a logical vector with a value for
each of its columns determining if it contains solar data or not. The dataframe must
have a date column named "date" as its last column. This can be added using the add_date
function. The function can be instructed to look for data in W/m2 or KW/m2 through the
W_or_KW parameter. Data is determined to be solar data if it passes 3 tests:
\enumerate{
\item data is between 0 and 1 for KW/m2, 0 and 1000 for W/m2
\item All winter data reads 0
\item All summer data fits a sine curve with period of 1 year
Data just before or after winter is difficult to analyse. If more than 80\% of data is
within this range, the file path is written to file in output_dir for manual inspection
}
}
