% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_plot_col_names.R
\name{standardise_plot_col_names}
\alias{standardise_plot_col_names}
\title{Standardise plot column names}
\usage{
standardise_plot_col_names(dataframe, output_dir, file_path)
}
\arguments{
\item{dataframe}{The dataframe which is to have its column names standardised}

\item{output_dir}{The directory to which warning text files may be produced}

\item{file_path}{The file_path which will be used to identify the data in the warning
text files}
}
\description{
A function which accepts a dataframe, and returns it with applicable column names
standardised to:
site
plot
otc_treatment
snow_treatment
fert_treatment
co2_plot
The output_dir and file_path of the original data are utilised by the find_treatment_cols
function to produce a warning file if it has to split a column containing a mix of snow and
otc treatment
}
