% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_plot_number.R
\name{standardise_plot_number}
\alias{standardise_plot_number}
\title{Standardise Plot Number}
\usage{
standardise_plot_number(dataframe)
}
\arguments{
\item{dataframe}{A dataframe that should have its plot numbers standardised}
}
\description{
A function which accepts a dataframe and returns it having standardised it's plot numbering.
First, if a column named sect (section) is present containing letters A-C, then the section is
appended to the plot number, before removing sect column. This is relevant for the BS_Annex sites.
At this point, if A cassiope or Willow site contains A-C in its plot number, then it is ensured the
site is named Annex_BS.
Next, any plot number containing A, R, or C and not belonging to an Annex or Farm site, has
"addition", "removal", or "control" added to its snow_treatment column. Finally, any plot not
belonging to an Annex, Farm, or Dome site has any letters removed from its plot number. Column name
standardisation should be performed before using this function. The required column names are:
"site"
"plot"
"sect"
"otc_treatment"
"snow_treatment" - This column will be added if it is not already present - however,
find_treatment_cols should be applied to the dataframe before the standardisation of plot number.
}
