% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_sites.R
\name{standardise_sites}
\alias{standardise_sites}
\title{Standardise Sites}
\usage{
standardise_sites(dataframe)
}
\arguments{
\item{dataframe}{The dataframe, which must contain a column named "site", which is to have it's
site names standardised}
}
\description{
A function which accepts a dataframe, standardises the names found within the column headed
"site", and returns that dataframe. The sites which are currently attempted to be standardised are:
Sax_opp,
Fert,
Beach_Ridge,
Cassiope,
Willow,
Meadow,
Dryas,
Annex_Willow,
Annex_Cassiope,
Annex_BS_Willow,
Annex_BS_Cassiope,
Annex_Migration
}
