% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_treatment.R
\name{standardise_treatment}
\alias{standardise_treatment}
\title{Standardise Treatment}
\usage{
standardise_treatment(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing columns named "site", "plot", and at least one column
with name matching regex "otc|treat" containing treatment data which is to be standardised}

\item{file_path}{A file path for the original data. Used in the output of a warning message if
snow and otc treatment data could not be clearly separated.}

\item{output_dir}{The directory that a warning file will be written to if snow and otc treatment
data can not be clearly separated.}
}
\description{
A function which accepts a dataframe and returns it with the contents of its treatment columns
standardised. find_treat_cols should be applied to the dataframe before treatment is standardised.
}
\details{
For each treatment column present, each term is examined for the presence of "CO2" in its name.
If "CO2" is part of the treatment descriptor then, "Y" is added to the column "co2_plot" (this column
is created if no CO2 plot already exists, or renamed if a CO2 column does exist with a different name).
If an annex_BS site has its plot letters ABC stored within the treatment column this is moved to
the plot number column and removed from treatment.
Finally, the treatment descriptors are renamed so that they are all written the same. The accepted
terms are:
OTC,
cover
control
addition
removal
}
