% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatments_from_plot_id.R
\name{treatments_from_plot_id}
\alias{treatments_from_plot_id}
\title{Clean treatment columns by looking up plot IDs}
\usage{
treatments_from_plot_id(dataframe)
}
\arguments{
\item{dataframe}{The dataframe that is to have its treatment data completed based on
the plot_ids it already contains and the corresponding treatment info for those plots found
in tundra::plot_names}
}
\description{
A function which accepts a dataframe with plot IDs and returns a dataframe with all
treatment columns (OTC, snow and fert), as well as the site and plot data completed based on
the information available for that plot in tundra::plot_names. If the plot ID can not be
identified then the treatment info is unchanged. The input dataframe must be in the format
produced by the standardise_plots function, in that it must have columns named "plot_id",
"otc_treatment", "snow_treatment", "fert_treatment", "site", and "plot".
}
