% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year_from_string.R
\name{year_extractor}
\alias{year_extractor}
\title{extract year from string}
\usage{
year_extractor(string, date_format = "all", breaks = FALSE, unique = TRUE)
}
\arguments{
\item{string}{A character vector that is to be examined for years}

\item{date_format}{which date formats will be checked. One of "all", "ymd" (year-month-day),
"dmy" (day-month-year), or "mdy" (month-day-year). Defaults to all. "four" only checks for
4 digit years}

\item{breaks}{Should results be restricted to years encapsulated by word boundaries}

\item{unique}{Should result return all unique hits or a full list for counting occurrences}
}
\description{
A function which accepts a character vector and returns all unique instances within that string
that match a year format. If no years are found then it returns NULL. Accepted formats include:
4 digits not directly next to any other digits
8 digits fitting year-month-day, day-month-year, or month-day-year formats.
Certain dates will throw multiple years as they can be read in different formats, for example
20062005 will return 2006 and 2005 because it fits both 20/jun/2005 and 20/may/2006. This can be
prevented by selecting a date format, which ensures all dates are read in only one way. Regardless
the date format, 4 digit years not directly touching other digits will always be returned. To further
prevent miss-atributation, only years 1980-2029 will be returned. Setting breaks to TRUE means that
only years isolated from other words are returned, for example "summer2012" would not produce a hit
but "summer_2012" or simply "2012" would. Setting unique to FALSE means all hits are returned, which
is useful for counting them. WARNING - if unique is set to FALSE, and a specific date format is not
selected, then extra hits may be generated, for example 01022003 will produce two hits, one as first
of February and a second as the second of January.
}
