% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_identifier.R
\name{year_identifier}
\alias{year_identifier}
\title{Identify which column of a dataframe contains year data}
\usage{
year_identifier(dataframe)
}
\arguments{
\item{dataframe}{A dataframe which is to be analysed for the presence of a Year column}
}
\description{
A function which analyses every column in dataframe and returns a logical vector where
the value true indicates the corresponding column contains a list of years.
A column is identified as containing year data if, ignoring all 6999 values, NAs,
non-numeric data, and the 1st and last quartiles, the data ranges from 1980-2030
and there are at least 10 data points for every unique value identified.
}
