% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_parameter.R
\name{parameter_present}
\alias{parameter_present}
\title{Is parameter present?}
\usage{
parameter_present(dataframe, extractor_function, col_name, ...)
}
\arguments{
\item{dataframe}{A dataframe to be examined}

\item{extractor_function}{A function which can search a string, or series of strings for the data
in question and return all examples it finds.}

\item{col_name}{A character vector giving the name of the column to search for}
}
\description{
A function which accepts a dataframe, a column name, and an extractor function:
A function which uses regular expressions to locate instances of data, and returns
all matching examples of the data it finds. The function checks if the dataframe contains
a column matching the given column name, and if it does it checks that at least 40\% of the
data results in a hit in the extractor function. If so, it returns TRUE, else FALSE.
}
