% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_extractor.R
\name{site_extractor}
\alias{site_extractor}
\title{extract site from string}
\usage{
site_extractor(string, unique = TRUE)
}
\arguments{
\item{string}{A character vector that is to be examined for sites}

\item{unique}{Should result return all unique hits or a full list for counting occurrences}
}
\description{
A function which accepts a character vector and returns all unique instances within that string
that match a site name. If no sites are found then it returns NULL. Setting unique to FALSE means
all hits are returned, which is useful for counting them. If multiple sites are present within one
string then an error is returned
}
